+++ hydromt/data_adapter/utils.py
@@ -1,90 +1,24 @@
 """Utility functions for data adapters."""
-import os
-from logging import Logger
-from os.path import isdir, join
-from pathlib import Path
-from typing import List, Optional, Union
+
+from logging import Logger, getLogger
+from typing import TYPE_CHECKING, Dict, Optional, Union, cast
 
 import geopandas as gpd
 import numpy as np
 import pandas as pd
 import xarray as xr
 
-from hydromt._typing import Data
+from hydromt._typing.type_def import TimeRange, Variables
 
+if TYPE_CHECKING:
+    from hydromt.data_source.data_source import SourceMetadata
 
-def netcdf_writer(
-    obj: Union[xr.Dataset, xr.DataArray],
-    data_root: Union[str, Path],
-    data_name: str,
-    variables: Optional[List[str]] = None,
-    encoder: str = "zlib",
-) -> str:
-    """Utiliy function for writing a xarray dataset/data array to a netcdf file.
-
-    Parameters
-    ----------
-    obj : xr.Dataset | xr.DataArray
-        Dataset.
-    data_root : str | Path
-        root to write the data to.
-    data_name : str
-        filename to write to.
-    variables : Optional[List[str]]
-        list of dataset variables to write, by default None
-
-    Returns
-    -------
-    fn_out: str
-        Absolute path to output file
-    """
-    dvars = [obj.name] if isinstance(obj, xr.DataArray) else obj.data_vars
-    if variables is None:
-        encoding = {k: {encoder: True} for k in dvars}
-        fn_out = join(data_root, f"{data_name}.nc")
-        obj.to_netcdf(fn_out, encoding=encoding)
-    else:  # save per variable
-        if not isdir(join(data_root, data_name)):
-            os.makedirs(join(data_root, data_name))
-        for var in dvars:
-            fn_out = join(data_root, data_name, f"{var}.nc")
-            obj[var].to_netcdf(fn_out, encoding={var: {encoder: True}})
-        fn_out = join(data_root, data_name, "{variable}.nc")
-    return fn_out
-
-
-def zarr_writer(
-    obj: Union[xr.Dataset, xr.DataArray],
-    data_root: Union[str, Path],
-    data_name: str,
-    **kwargs,
-) -> str:
-    """Utiliy function for writing a xarray dataset/data array to a netcdf file.
-
-    Parameters
-    ----------
-    obj : xr.Dataset | xr.DataArray
-        Dataset.
-    data_root : str | Path
-        root to write the data to.
-    data_name : str
-        filename to write to.
-
-    Returns
-    -------
-    fn_out: str
-        Absolute path to output file
-    """
-    fn_out = join(data_root, f"{data_name}.zarr")
-    if isinstance(obj, xr.DataArray):
-        obj = obj.to_dataset()
-    obj.to_zarr(fn_out, **kwargs)
-    return fn_out
+logger = getLogger(__name__)
 
 
 def shift_dataset_time(
-    dt: int, ds: Data, logger: Logger, time_unit: str = "s"
-) -> xr.Dataset:
+    dt: int, ds: Optional[xr.Dataset], logger: Logger, time_unit: str = "s"
+) -> Optional[xr.Dataset]:
     """Shifts time of a xarray dataset.
 
     Parameters
@@ -101,6 +35,9 @@ def shift_dataset_time(
     xr.Dataset
         time shifted dataset
     """
+    if ds is None:
+        return None
+
     if (
         dt != 0
         and "time" in ds.dims
@@ -124,3 +61,108 @@ def has_no_data(
         return all([v.size == 0 for v in data.data_vars.values()])
     else:
         return len(data) == 0
+
+
+def _single_var_as_array(
+    maybe_ds: Optional[xr.Dataset],
+    single_var_as_array: bool,
+    variable_name: Optional[Variables] = None,
+) -> Optional[xr.Dataset]:
+    if maybe_ds is None:
+        return None
+    else:
+        ds = maybe_ds
+    # return data array if single variable dataset
+    dvars = list(ds.data_vars.keys())
+    if single_var_as_array and len(dvars) == 1:
+        da = ds[dvars[0]]
+        if isinstance(variable_name, list) and len(variable_name) == 1:
+            da.name = variable_name[0]
+        elif isinstance(variable_name, str):
+            da.name = variable_name
+        return da
+    else:
+        return ds
+
+
+def _set_vector_nodata(
+    ds: Optional[xr.Dataset], metadata: "SourceMetadata"
+) -> Optional[xr.Dataset]:
+    if ds is None:
+        return None
+    if metadata.nodata is not None:
+        if not isinstance(metadata.nodata, dict):
+            nodata = {k: metadata.nodata for k in ds.data_vars.keys()}
+        else:
+            nodata = metadata.nodata
+        for k in ds.data_vars:
+            mv = nodata.get(k, None)
+            if mv is not None and ds[k].vector.nodata is None:
+                ds[k].vector.set_nodata(mv)
+    return ds
+
+
+def _set_raster_nodata(
+    ds: Optional[xr.Dataset], metadata: "SourceMetadata"
+) -> Optional[xr.Dataset]:
+    if ds is None:
+        return None
+    if metadata.nodata is not None:
+        if not isinstance(metadata.nodata, dict):
+            nodata = {k: metadata.nodata for k in ds.data_vars.keys()}
+        else:
+            nodata = metadata.nodata
+        for k in ds.data_vars:
+            mv = nodata.get(k, None)
+            if mv is not None and ds[k].raster.nodata is None:
+                ds[k].raster.set_nodata(mv)
+    return ds
+
+
+def _slice_temporal_dimension(
+    ds: Optional[xr.Dataset],
+    time_range: Optional[TimeRange],
+    logger: Logger = logger,
+    # TODO: https://github.com/Deltares/hydromt/issues/802
+) -> Optional[xr.Dataset]:
+    if ds is None:
+        return None
+    else:
+        if (
+            "time" in ds.dims
+            and ds["time"].size > 1
+            and np.issubdtype(ds["time"].dtype, np.datetime64)
+        ):
+            logger.debug(f"Slicing time dim {time_range}")
+            ds = ds.sel(time=slice(*time_range))
+        if has_no_data(ds):
+            return None
+        else:
+            return ds
+
+
+def _set_metadata(
+    ds: Optional[xr.Dataset], metadata: "SourceMetadata"
+) -> Optional[xr.Dataset]:
+    if ds is None:
+        return None
+    elif metadata.attrs:
+        if isinstance(ds, xr.DataArray):
+            name = cast(str, ds.name)
+            ds.attrs.update(metadata.attrs[name])
+        else:
+            for k in metadata.attrs:
+                ds[k].attrs.update(metadata.attrs[k])
+
+    ds.attrs.update(metadata)
+    return ds
+
+
+def _rename_vars(
+    ds: Optional[xr.Dataset], rename: Dict[str, str]
+) -> Optional[xr.Dataset]:
+    if ds is None:
+        return None
+    rm = {k: v for k, v in rename.items() if k in ds}
+    ds = ds.rename(rm)
+    return ds